+++ CyRK/cy/common.pyx
@@ -1,0 +1,3 @@
+# distutils: language = c++
+# cython: boundscheck=False, wraparound=False, nonecheck=False, cdivision=True, initializedcheck=False
+MAX_SIZE = MAX_SIZET_SIZE