+++ CyRK/cy/cysolver.hpp
@@ -7,8 +7,10 @@
 #include <functional>
 #include <memory>
 
-#include "common.hpp"
+#include "c_common.hpp"
 #include "cy_array.hpp"
+#include "c_events.hpp"
+#include "c_brentq.hpp"
 
 // !!!
 // Comment the following 
@@ -92,6 +94,10 @@ struct ProblemConfig {
     // Flags
     bool initialized = false;
 
+    // Event data
+    bool check_events = false;
+    std::vector<Event> events_vec = std::vector<Event>();
+
     // Solver specific configurations can be added below via overloading the class.
 
     virtual ~ProblemConfig() {};
@@ -115,7 +121,8 @@ struct ProblemConfig {
         size_t max_ram_MB_,
         PreEvalFunc pre_eval_func_,
         bool capture_dense_output_,
-        bool force_retain_solver_);
+        bool force_retain_solver_,
+        std::vector<Event>& events_vec_);
 
     // Helper functions
     void update_properties(
@@ -137,7 +144,8 @@ struct ProblemConfig {
         size_t max_ram_MB_,
         PreEvalFunc pre_eval_func_,
         bool capture_dense_output_,
-        bool force_retain_solver_
+        bool force_retain_solver_,
+        std::vector<Event>& events_vec_
     );
     virtual void initialize();
     virtual void update_properties_from_config(ProblemConfig* new_config_ptr);
@@ -165,6 +173,7 @@ protected:
     virtual void p_step_implementation() noexcept;
     inline void p_cy_diffeq() noexcept;
     virtual void p_calc_first_step_size() noexcept;
+    CyrkErrorCodes p_check_events() noexcept;
 
 public:
     CySolverBase();
@@ -236,6 +245,7 @@ protected:
     bool setup_called      = false;
     bool error_flag        = false;
     bool capture_extra     = false;
+    bool check_events_flag = false;
 
     // Dependent variable pointers
     double* y0_ptr = nullptr;
@@ -248,10 +258,17 @@ protected:
     double* y_tmp_ptr    = nullptr;
     double* y_interp_ptr = nullptr;
     // For dy, both the dy/dt and any extra outputs are stored. So the maximum size is `num_y` + `num_extra`
-    std::vector<double> dy_holder_vec = std::vector<double>(PRE_ALLOC_NUMY * 3);
+    std::vector<double> dy_holder_vec = std::vector<double>(PRE_ALLOC_NUMY * 4);
     double* dy_old_ptr = nullptr;
     //double* dy_now_ptr = nullptr; // This needs to be public
-    double* dy_tmp_ptr = nullptr;
+    double* dy_tmp_ptr  = nullptr;
+    double* dy_tmp2_ptr = nullptr;
+
+    // Event data holder
+    OptimizeInfo root_finder_data      = OptimizeInfo();
+    std::vector<double> event_data_vec = std::vector<double>();  // Equivalent to SciPy's "g" array.
+    double* event_checks_old_ptr       = nullptr;
+    double termination_root            = 0.0;
 
     // Integration step information
     size_t max_num_steps = 0;
@@ -272,9 +289,11 @@ public:
     size_t num_y       = 0;
     size_t num_extra   = 0;
     size_t num_dy      = 0;
+    size_t num_events  = 0;
     double t_old       = 0.0;
     double t_now       = 0.0;
     double* y_old_ptr  = nullptr;
     double* y_now_ptr  = nullptr;
     double* dy_now_ptr = nullptr;
+    std::vector<size_t> active_event_indices_vec = std::vector<size_t>();
 };