+++ CyRK/cy/cysolver_test.pyx
@@ -10,7 +10,9 @@ from libcpp.vector cimport vector
 
 cdef double d_NAN = numeric_limits[double].quiet_NaN()
 
-from CyRK.cy.cysolver_api cimport cysolve_ivp_noreturn, cysolve_ivp, WrapCySolverResult, DiffeqFuncType, MAX_STEP, CySolveOutput, ODEMethod, CySolverResult
+from CyRK.cy.common cimport DiffeqFuncType, MAX_STEP
+from CyRK.cy.cysolver_api cimport cysolve_ivp_noreturn, cysolve_ivp, WrapCySolverResult, CySolveOutput, ODEMethod, CySolverResult
+from CyRK.cy.events cimport Event
 
 import numpy as np
 cimport numpy as np
@@ -219,7 +221,6 @@ cdef void pendulum_preeval_diffeq(double* dy_ptr, double t, double* y_ptr, char*
     dy_ptr[0] = y1
     dy_ptr[1] = pre_eval_storage_ptr[1] * sin(y0) + pre_eval_storage_ptr[2] * pre_eval_storage_ptr[0]
 
-
 def cy_extra_output_tester():
 
     cdef double t_start = 0.0
@@ -282,6 +283,54 @@ def cy_extra_output_tester():
 
     return True
 
+# Event fuinctions for baseline diffeq
+cdef double baseline_diffeq_event1_check(double t, double* y, char* args) noexcept nogil:
+    if y[0] > 50.0:
+        return 0.0
+    return 1.0
+
+cdef double baseline_diffeq_event2_check(double t, double* y, char* args) noexcept nogil:
+    if y[1] < 50.0:
+        return 0.0
+    return 1.0
+
+cdef double baseline_diffeq_event3_check(double t, double* y, char* args) noexcept nogil:
+    if t > 5.0:
+        return 0.0
+    else:
+        return 1.0
+
+cdef vector[Event] baseline_diffeq_events = vector[Event]()
+# Default constructor will handle direction and max num allowed (no termination in this case)
+baseline_diffeq_events.emplace_back(baseline_diffeq_event1_check)
+baseline_diffeq_events.emplace_back(baseline_diffeq_event2_check)
+baseline_diffeq_events.emplace_back(baseline_diffeq_event3_check)
+
+
+# Event functions for pendulum diffeq
+cdef double pendulum_diffeq_event1_check(double t, double* y, char* args) noexcept nogil:
+    if y[0] > 50.0:
+        return 0.0
+    return 1.0
+
+cdef double pendulum_diffeq_event2_check(double t, double* y, char* args) noexcept nogil:
+    if y[1] < 50.0:
+        return 0.0
+    return 1.0
+
+cdef double pendulum_diffeq_event3_check(double t, double* y, char* args) noexcept nogil:
+    if t > 5.0:
+        return 0.0
+    else:
+        return 1.0
+
+cdef vector[Event] pendulum_diffeq_events = vector[Event]()
+# Default constructor will handle direction and max num allowed (no termination in this case)
+pendulum_diffeq_events.emplace_back(pendulum_diffeq_event1_check)
+pendulum_diffeq_events.emplace_back(pendulum_diffeq_event2_check)
+pendulum_diffeq_events.emplace_back(pendulum_diffeq_event3_check)
+
+
 def cytester(
         int diffeq_number,
         tuple t_span = None,
@@ -292,6 +341,7 @@ def cytester(
         size_t max_num_steps = 0,
         size_t max_ram_MB = 2000,
         bint dense_output = False,
+        bint use_events = False,
         double[::1] t_eval = None,
         double rtol = 1.0e-3, 
         double atol = 1.0e-6,
@@ -355,6 +405,15 @@ def cytester(
         pre_eval_func = pendulum_preeval_func
     else:
         raise NotImplementedError
+    
+    cdef vector[Event] events_vec = vector[Event]()
+    if use_events:
+        if diffeq_number == 0:
+            events_vec = baseline_diffeq_events
+        elif diffeq_number == 6:
+            events_vec = pendulum_diffeq_events
+        else:
+            raise NotImplementedError
 
     # Set up additional argument information
     cdef vector[char] args_vec = vector[char]()
@@ -492,8 +551,6 @@ def cytester(
             # This tester assumes that the args input is an array of doubles.
             args_vec.resize(args.size * sizeof(double))
             memcpy(args_vec.data(), &args[0], sizeof(double) * args.size)
-            args_ptr     = <char*>&args[0]
-            size_of_args = sizeof(double) * args.size
         else:
             args_vec.resize(0)
 
@@ -542,6 +599,7 @@ def cytester(
         dense_output = dense_output,
         t_eval_vec = t_eval_vec,
         pre_eval_func = pre_eval_func,
+        events_vec = events_vec,
         rtols_vec = rtols_vec,
         atols_vec = atols_vec,
         max_step = max_step,