+++ CyRK/cy/events.hpp
@@ -1,3 +1,0 @@
-#pragma once
-
-typedef double (*EventFunc)(double, char*);